Move checkout helpers outside Netlify functions
