+++ netlify/functions/create-checkout-session.test.js
@@ -1,7 +1,7 @@
 const assert = require('node:assert/strict');
 const test = require('node:test');
 
-const { CLASSES } = require('./_checkout/catalog');
+const { CLASSES } = require('../lib/checkout/catalog');
 const { _test } = require('./create-checkout-session');
 
 const ORIGIN = 'https://www.italianschoolsd.com';
