+++ netlify/lib/checkout/catalog.js
